requirements.txt
streamlit
joblib
Pillow
scikit-learn
numpy
